requirements.txt
PyQt5
pandas
numpy
yfinance
alpha_vantage
python-telegram-bot
scikit-learn
matplotlib
openpyxl       
reportlab      
pytest
